Add Provider and update reducer, action and store files
