+++ .eslintrc.json
@@ -22,7 +22,8 @@
     "jsx-a11y/click-events-have-key-events":"off",
     "jsx-a11y/no-static-element-interactions":"off",
     "react/no-access-state-in-setstate":"off",
-    "react/no-unused-state":"off"
+    "react/no-unused-state":"off",
+    "no-underscore-dangle":"off"
   },
   "ignorePatterns": [
     "dist/",
