+++ src/redux/book/booksReducer.js
@@ -1,13 +1,16 @@
+/* eslint-disable */
+import {addBook}  from "./actions/bookActions";
+import {removeBook}  from "./actions/bookActions";
 const initialState = [];
-const bookReducer=(state=initialState, action)=>{
-    switch(action.type){
-        case ADD_BOOK:
-            return state;
-        case REMOVE_BOOK:
-            return state;
-        default:
-            return state;
-    }
-}
+const bookReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case addBook.type:
+      return state;
+    case removeBook.type:
+      return state;
+    default:
+      return state;
+  }
+};
 
-export default bookReducer
+export default bookReducer;
